Revert "[builder] Add builder module -- formerly ci/build.py (#6376)"
This reverts commit 7c998fe9106b13abc1f62414f0bd42ecf638f28c.
