+++ build.yaml
@@ -157,10 +157,10 @@ steps:
    dependsOn:
      - base_image
  - kind: buildImage
-   name: builder_utils_image
-   dockerFile: builder/Dockerfile
-   contextPath: builder/
-   publishAs: builder-utils
+   name: ci_utils_image
+   dockerFile: ci/Dockerfile.ci-utils
+   contextPath: ci
+   publishAs: ci-utils
    dependsOn:
      - base_image
  - kind: runImage
@@ -458,7 +458,7 @@ steps:
    dependsOn:
     - default_ns
     - ci_image
-    - builder_utils_image
+    - ci_utils_image
     - create_accounts
     - ci_database
     - create_ci_tables
