--- builder/.dockerignore
@@ -1,3 +1,0 @@
-*~
-*.pyc
-**/__pycache__
