--- builder/.gitignore
@@ -1,8 +1,0 @@
-/builder.log
-/Dockerfile.out
-*.pyc
-**/__pycache__
-flycheck_*.py
-*_flymake.py
-/flake8-stmp
-/pylint-stmp
