--- builder/builder/__init__.py
@@ -1,6 +1,0 @@
-from .builder import BuildConfiguration, Code
-
-__all__ = [
-    'BuildConfiguration',
-    'Code'
-]
