--- builder/builder/utils.py
@@ -1,13 +1,0 @@
-import string
-import secrets
-
-
-def generate_token(size=12):
-    assert size > 0
-    alpha = string.ascii_lowercase
-    alnum = string.ascii_lowercase + string.digits
-    return secrets.choice(alpha) + ''.join([secrets.choice(alnum) for _ in range(size - 1)])
-
-
-def flatten(xxs):
-    return [x for xs in xxs for x in xs]
