--- builder/setup.cfg
@@ -1,3 +1,0 @@
-[flake8]
-# E501 line too long, W504 no newlines after binary operators
-ignore=E501,W504
