+++ ci/Dockerfile
@@ -10,11 +10,6 @@ COPY batch_client/batch_client /batch_client/batch_client
 RUN pip3 install --no-cache-dir /batch_client && \
   rm -rf /batch_client
 
-COPY builder/setup.py /builder/
-COPY builder/builder /builder/builder
-RUN python3 -m pip install --no-cache-dir /builder && \
-  rm -rf /builder
-
 COPY ci/requirements.txt .
 RUN python3 -m pip install --no-cache-dir -U -r requirements.txt
