+++ ci/ci-utils-requirements.txt
@@ -1,0 +1,1 @@
+jinja2
