+++ ci/ci/constants.py
@@ -3,8 +3,6 @@ import gear.auth as hj
 
 GITHUB_CLONE_URL = 'https://github.com/'
 
-SELF_HOSTNAME = os.environ['HAIL_SELF_HOSTNAME']
-
 with open(os.environ['HAIL_TOKEN_FILE']) as f:
     userdata = hj.JWTClient.unsafe_decode(f.read())
     BUCKET = f'gs://{userdata["bucket_name"]}'
