+++ ci/ci/environment.py
@@ -1,0 +1,7 @@
+import os
+
+GCP_PROJECT = os.environ['HAIL_GCP_PROJECT']
+DOMAIN = os.environ['HAIL_DOMAIN']
+IP = os.environ['HAIL_IP']
+CI_UTILS_IMAGE = os.environ['HAIL_CI_UTILS_IMAGE']
+SELF_HOSTNAME = os.environ['HAIL_SELF_HOSTNAME']
