+++ ci/ci/github.py
@@ -7,12 +7,11 @@ import concurrent.futures
 import aiohttp
 import gidgethub
 import humanize
-
-from builder import BuildConfiguration, Code
-
 from .log import log
-from .constants import GITHUB_CLONE_URL, AUTHORIZED_USERS, SELF_HOSTNAME
+from .constants import GITHUB_CLONE_URL, AUTHORIZED_USERS
+from .environment import SELF_HOSTNAME
 from .utils import check_shell, check_shell_output
+from .build import BuildConfiguration, Code
 
 repos_lock = asyncio.Lock()
