+++ ci/ci/utils.py
@@ -1,3 +1,5 @@
+import string
+import secrets
 import asyncio
 
 
@@ -27,3 +29,14 @@ async def check_shell_output(script):
     if proc.returncode != 0:
         raise CalledProcessError(script, proc.returncode)
     return outerr
+
+
+def generate_token(size=12):
+    assert size > 0
+    alpha = string.ascii_lowercase
+    alnum = string.ascii_lowercase + string.digits
+    return secrets.choice(alpha) + ''.join([secrets.choice(alnum) for _ in range(size - 1)])
+
+
+def flatten(xxs):
+    return [x for xs in xxs for x in xs]
